Rename vmgraph package to kvgraph
