+++ pkg/plugin/vmi_restore_item_action.go
@@ -29,7 +29,7 @@ import (
 	"k8s.io/apimachinery/pkg/runtime"
 	kvcore "kubevirt.io/api/core/v1"
 
-	vmgraph "kubevirt.io/kubevirt-velero-plugin/pkg/util/graph"
+	"kubevirt.io/kubevirt-velero-plugin/pkg/util/kvgraph"
 )
 
 // VMIRestorePlugin is a VMI restore item action plugin for Velero (duh!)
@@ -91,7 +91,7 @@ func (p *VMIRestorePlugin) Execute(input *velero.RestoreItemActionExecuteInput) 
 	metadata.SetLabels(labels)
 
 	output := velero.NewRestoreItemActionExecuteOutput(input.Item)
-	output.AdditionalItems = vmgraph.NewVirtualMachineInstanceRestoreGraph(vmi)
+	output.AdditionalItems = kvgraph.NewVirtualMachineInstanceRestoreGraph(vmi)
 	return output, nil
 }
