+++ pkg/util/kvgraph/restore_graph_test.go
@@ -1,4 +1,4 @@
-package vmgraph
+package kvgraph
 
 import (
 	"testing"
